+++ src/main/java/com/getindata/connectors/http/internal/table/lookup/JavaNetHttpPollingClient.java
@@ -9,11 +9,9 @@ import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -61,7 +59,7 @@ public class JavaNetHttpPollingClient implements PollingClient {
     private final ObjectMapper objectMapper;
     private final HttpPostRequestCallback<HttpLookupSourceRequestEntry> httpPostRequestCallback;
     private final HttpLookupConfig options;
-    private final Set<Integer> ignoredErrorCodes;
+    private final HttpResponseChecker responseChecker;
     private final boolean continueOnError;
 
     public JavaNetHttpPollingClient(
@@ -77,17 +75,16 @@ public class JavaNetHttpPollingClient implements PollingClient {
         this.options = options;
         var config = options.getReadableConfig();
 
-        this.ignoredErrorCodes = HttpCodesParser.parse(config.get(SOURCE_LOOKUP_HTTP_IGNORED_RESPONSE_CODES));
+        var ignoreCodes = HttpCodesParser.parse(config.get(SOURCE_LOOKUP_HTTP_IGNORED_RESPONSE_CODES));
         var errorCodes = HttpCodesParser.parse(config.get(SOURCE_LOOKUP_HTTP_RETRY_CODES));
-        var successCodes = new HashSet<Integer>();
-        successCodes.addAll(HttpCodesParser.parse(config.get(SOURCE_LOOKUP_HTTP_SUCCESS_CODES)));
-        successCodes.addAll(ignoredErrorCodes);
+        var successCodes = HttpCodesParser.parse(config.get(SOURCE_LOOKUP_HTTP_SUCCESS_CODES));
         this.continueOnError = config.get(SOURCE_LOOKUP_CONTINUE_ON_ERROR);
+        this.responseChecker = new HttpResponseChecker(successCodes, errorCodes, ignoreCodes);
 
         this.httpClient = HttpClientWithRetry.builder()
                 .httpClient(httpClient)
                 .retryConfig(RetryConfigProvider.create(config))
-                .responseChecker(new HttpResponseChecker(successCodes, errorCodes))
+                .responseChecker(responseChecker)
                 .build();
     }
 
@@ -215,7 +212,7 @@ public class JavaNetHttpPollingClient implements PollingClient {
 
         log.debug("Received status code [{}] for RestTableSource request with Server response body [{}] ",
                 response.statusCode(), responseBody);
-        if (!isError && (StringUtils.isNullOrWhitespaceOnly(responseBody) || ignoreResponse(response))) {
+        if (!isError && (StringUtils.isNullOrWhitespaceOnly(responseBody) || responseChecker.isIgnoreCode(response))) {
             return HttpRowDataWrapper.builder()
                     .data(Collections.emptyList())
                     .httpCompletionState(HttpCompletionState.SUCCESS)
@@ -292,9 +289,5 @@ public class JavaNetHttpPollingClient implements PollingClient {
             }
         }
         return result;
-    }
-
-    private boolean ignoreResponse(HttpResponse<?> response) {
-        return ignoredErrorCodes.contains(response.statusCode());
     }
 }