+++ src/test/java/com/getindata/connectors/http/internal/table/sink/HttpDynamicTableSinkFactoryTest.java
@@ -111,4 +111,24 @@ public class HttpDynamicTableSinkFactoryTest {
         assertThrows(ValidationException.class,
             () -> tEnv.executeSql("INSERT INTO http VALUES (1)").await());
     }
+
+    @Test
+    public void exactlyOnceDeliveryGuaranteeNotSupported() {
+        final String exactlyOnceCreateSql =
+            String.format(
+                "CREATE TABLE http (\n"
+                    + "  id bigint\n"
+                    + ") with (\n"
+                    + "  'connector' = '%s',\n"
+                    + "  'url' = '%s',\n"
+                    + "  'format' = 'json',\n"
+                    + "  'sink.delivery-guarantee' = 'exactly-once'\n"
+                    + ")",
+                HttpDynamicTableSinkFactory.IDENTIFIER,
+                "http://localhost/"
+            );
+        tEnv.executeSql(exactlyOnceCreateSql);
+        assertThrows(ValidationException.class,
+            () -> tEnv.executeSql("INSERT INTO http VALUES (1)").await());
+    }
 }